new file tree and added consts where it's needed.
