+++ lib/src/menuPages/comunicados_page.dart
@@ -1,7 +1,5 @@
-// ignore_for_file: file_names
-
 import 'package:flutter/material.dart';
-import 'package:mermas_digitais_app/models/postModel.dart';
+import 'package:mermas_digitais_app/src/models/post_model.dart';
 
 class ComunicadosPage extends StatefulWidget {
   const ComunicadosPage({super.key});
