+++ lib/src/menuPages/frequencia_page.dart
@@ -5,7 +5,7 @@ import 'package:firebase_auth/firebase_auth.dart';
 import 'package:firebase_storage/firebase_storage.dart';
 import 'package:flutter/material.dart';
 import 'package:iconsax/iconsax.dart';
-import 'package:mermas_digitais_app/models/loadingWindow.dart';
+import 'package:mermas_digitais_app/src/models/loading_window.dart';
 import 'package:percent_indicator/percent_indicator.dart';
 
 class FrequenciaPage extends StatefulWidget {
