+++ lib/src/models/oficina_model.dart
