+++ lib/src/models/post_model.dart
